+++ instalar.sh
@@ -13,7 +13,7 @@
 #
 set -eu
 
-VERSION="1.1.0"
+VERSION="1.2.0"
 REPO="all-lopezg/trawun"
 RAMIFICACION="main"